- moved http directive generation logic to separate module.
