+++ src/core/generator/json/mod.rs
@@ -2,6 +2,7 @@ mod field_base_url_generator;
 mod query_generator;
 mod schema_generator;
 mod types_generator;
+mod http_directive_generator;
 mod url_utils;
 
 pub use field_base_url_generator::FieldBaseUrlGenerator;
